requirements.txt
streamlit==1.50.0
requests==2.32.5
openai==2.9.0
python-dotenv==1.2.1
markdown-it-py==3.0.0

